New Run/DQM Page

+++ uploader/upload_step2.sh
@@ -11,7 +11,7 @@
 # Initialize Variables #
 ########################
 
-run_num=$1
+#run_num=$1
 uploads=/home/django/testing_database_hb/media/uploads
 cardDir=$uploads/run_control/cards    # temporary location of the qie card qc data
 #runDir=/home/django/testing_database_hb/media/uploads/run_control/run${run_num}_output    # location of run control tests
@@ -34,50 +34,53 @@ rm -f ${logLoc}*.log
 echo -e "${STATUS}Initial data set"
 echo ""
 
+if [ "$1" -eq 3 ] || [ "$1" -eq 1 ]; then
 ################################
 # Upload Quality Control Tests #
 ################################
-echo -e "${STATUS}Uploading Quality Control Tests"
+    echo -e "${STATUS}Uploading Quality Control Tests"
 
-# Check if there are any cards to upload
-if ls $cardDir/0x* &> /dev/null; then
-    for dir in $cardDir/0x*; do
-        [ -d "${dir}" ] || continue
-        qieuid="$(basename "${dir}")"    # list of uid directories
-        echo -e "    ${ACTION}Processing Card with UID: ${DEF}${qieuid}"
-        uidjsonFile=${dir}/${qieuid}_QC.json
-        python $scriptLoc/upload_qc.py $uidjsonFile 2> $logLoc${qieuid}.log
+    # Check if there are any cards to upload
+    if ls $cardDir/0x* &> /dev/null; then
+        for dir in $cardDir/0x*; do
+            [ -d "${dir}" ] || continue
+            qieuid="$(basename "${dir}")"    # list of uid directories
+            echo -e "    ${ACTION}Processing Card with UID: ${DEF}${qieuid}"
+            uidjsonFile=${dir}/${qieuid}_QC.json
+            python $scriptLoc/upload_qc.py $uidjsonFile 2> $logLoc${qieuid}.log
         
-        # Erase log files if there was no error
-        if [ $? -eq 0 ]; then
-            echo -e "    ${SUCCESS}Card Uploaded Succesfully"
-            rm $logLoc${qieuid}.log 
-        else
-            echo -e "    ${FAIL}ERROR: ${DEF} See log file: ${logLoc}${qieuid}.log"
-        fi
-    done
-    echo -e "${STATUS}No More Cards to Upload. Check log file if there was an error."
-else
-    echo -e "${FAIL}No Quality Control Data Found${DEF}"
+            # Erase log files if there was no error
+            if [ $? -eq 0 ]; then
+                echo -e "    ${SUCCESS}Card Uploaded Succesfully"
+                rm $logLoc${qieuid}.log 
+            else
+                echo -e "    ${FAIL}ERROR: ${DEF} See log file: ${logLoc}${qieuid}.log"
+            fi
+        done
+        echo -e "${STATUS}No More Cards to Upload. Check log file if there was an error.${DEF}"
+    else
+        echo -e "${FAIL}No Quality Control Data Found${DEF}"
+    fi
+    echo -e ""
 fi
-echo -e ""
 
+if [ "$1" -eq 3 ] || [ "$1" -eq 2 ]; then
 #########################
 # Upload Register Tests #
 #########################
 
-echo -e "${STATUS}Uploading Register Tests${DEF}"
+    echo -e "${STATUS}Uploading Register Tests${DEF}"
 
-if ls $regCardDir/0x* &> /dev/null; then
-    for dir in $regCardDir/0x*; do
-        [ -d "${dir}" ] || continue
-        qieuid="$(basename "${dir}")"
-        echo -e "    ${ACTION}Processing Card with UID: ${DEF}${qieuid}"
-    done
-else
-    echo -e "${FAIL}No Register Test Data Found${DEF}"
+    if ls $regCardDir/0x* &> /dev/null; then
+        for dir in $regCardDir/0x*; do
+            [ -d "${dir}" ] || continue
+            qieuid="$(basename "${dir}")"
+            echo -e "    ${ACTION}Processing Card with UID: ${DEF}${qieuid}"
+        done
+    else
+        echo -e "${FAIL}No Register Test Data Found${DEF}"
+    fi
 fi
-
